requirements.txt
kafka-python==2.0.2
pyspark==3.3.4
apache-airflow==2.7.2
streamlit==1.29.0
boto3==1.29.3
python-dotenv==1.0.0
apache-airflow-providers-apache-kafka==5.1.1
apache-airflow-providers-apache-spark==3.3.0
pandas==2.0.3
plotly==5.17.0
pyarrow==13.0.0
minio==7.2.3
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
scipy==1.10.1
fastparquet==2023.10.0
python-dateutil==2.8.2
psutil==5.9.6
jupyter==1.0.0
hdfs==3.6.0
kazoo==2.8.0
pyhive==0.6.1
thrift-sasl==0.4.4
impyla==0.17.0
structlog==23.1.0
elasticsearch==8.10.0
elasticsearch-dsl==8.1.1
prometheus-client==0.19.0
kafka-admin==0.1.11
kafka-python==2.0.2
confluent-kafka==2.0.2
kafka-connect-python==0.0.10
